backend/requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-cors-headers==3.10.0
django-polymorphic==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==9.3.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.8
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.1
typing-extensions==3.10.0.2
